Alt and Title Tags for 404 Pages
Addition of the Base Translation String Placeholders in the translations.yml, and the markdown tags within the image and link tags within the 404 pages within the /en and /nl folder to call the required translation string.

No content changes have been done in this commit. The alt and title tags have been left as are, if available.