Setting Translation Strings for Alt & Title Attributes for the Home Pages & Typo Fix for the Previous Commit on the 404 Pages
Addition of Translation Strings for the Homepage (index.md) for the alt and titles of the links. This change has been committed to all the language directories referenced in this commit.

Fixing a typo in how the translation strings were referenced in the 404 files, where the title attribute for the links was referred to as an alt initially.