Update to pers.md in EN & Replacement of Tel Linked Number with Translation String
In the process of updating the pers.md in the EN directory, it realised that the contact number is hardcoded in.

The contact number has been set as a translation string so that it has consistent uniformity in format across the pages where the tag is used, as well as that it is always linked (which was not prior to committing this change).

In the process, I have also replaced the occurances where the number is used with the translation string in the other language directories.