Government.nl Link in the Footer
If you click on the Government.nl link in the bottom right section of the footer, you end up on rijksoverheid.nl, so it is functionally correct.

However, you land on Dutch pages. When you scroll to the bottom and click on the switch to English button, you are then redirected to https://www.government.nl, which is in English.

But, furthermore you need to click on the text link 'Information about coronavirus COVID-19' to end up on the English section available for Covid information within the Government portal.

It would be optimal for the English version if the original link was set to https://www.government.nl/topics/coronavirus-covid-19.